+++ CHANGELOG.md
@@ -891,6 +891,10 @@
 * Fixes
   * Attach Javadocs and sources to Maven Central build.
 
+## Unreleased
+
+* Resolve issue where optional shipping parameters were treated as if they were required
+
 ## 3.0.0
 
 * Convert to AndroidX